backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python
scikit-image
scikit-learn
ultralytics
